refactor(server): change socket.io version, change from js to ts.
Bump socket.io version to latest (4.7.5). Migrate from v2 to v4.

Also add support for typescript and refactor from js to ts.
