--- server/router.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-router.get("/", (req, res) => {
-  res.send({ response: "Server is up and running." }).status(200);
-});
-
-module.exports = router;
