+++ server/router.ts
@@ -1,0 +1,9 @@
+import express, { Request, Response, Router } from "express";
+
+const router: Router = express.Router();
+
+router.get("/", (req: Request, res: Response) => {
+  res.send({ response: "Server is up and running." }).status(200);
+});
+
+export default router;
